Factor out install/uninstall and constants
